# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BattleScope API
  version: 2.0.0
  description: |
    BattleScope is a data intelligence platform that reconstructs and classifies battles in EVE Online
    by clustering related killmails from zKillboard.

    ## Key Features
    - Battle reconstruction from killmail clustering
    - Real-time killmail feed with Server-Sent Events
    - Entity name resolution via ESI API
    - Configurable ruleset filtering

    ## Data Types
    All EVE Online entity IDs (killmail, character, corporation, alliance, system, ship type) are
    transmitted as strings to support bigint values that exceed JavaScript's Number.MAX_SAFE_INTEGER.

  contact:
    name: BattleScope Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.battlescope.example.com
    description: Production server
...
paths:
  /battles:
    get:
      tags:
        - Battles
      summary: List battles
      description: |
        Returns a paginated list of battles with optional filtering by space type, system,
        alliance, corporation, character, or time range.
      operationId: listBattles
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/SpaceTypeParam'
        - $ref: '#/components/parameters/SystemIdParam'
        - $ref: '#/components/parameters/AllianceIdParam'
        - $ref: '#/components/parameters/CorpIdParam'
        - $ref: '#/components/parameters/CharacterIdParam'
        - $ref: '#/components/parameters/SinceParam'
        - $ref: '#/components/parameters/UntilParam'
      responses:
        '200':
          description: List of battles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /battles/{id}:
    get:
      tags:
        - Battles
      summary: Get battle details
      description: Returns detailed information about a specific battle including all killmails and participants
      operationId: getBattleById
      parameters:
        - name: id
          in: path
          required: true
          description: Battle UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Battle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /alliances/{id}/battles:
    get:
      tags:
        - Battles
      summary: List battles for an alliance
      description: Returns battles where the specified alliance participated
      operationId: listAllianceBattles
      parameters:
        - name: id
          in: path
          required: true
          description: Alliance ID
          schema:
            type: string
            pattern: '^\d+$'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/SpaceTypeParam'
        - $ref: '#/components/parameters/SinceParam'
        - $ref: '#/components/parameters/UntilParam'
      responses:
        '200':
          description: List of battles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /corporations/{id}/battles:
    get:
      tags:
        - Battles
      summary: List battles for a corporation
      description: Returns battles where the specified corporation participated
      operationId: listCorporationBattles
      parameters:
        - name: id
          in: path
          required: true
          description: Corporation ID
          schema:
            type: string
            pattern: '^\d+$'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/SpaceTypeParam'
        - $ref: '#/components/parameters/SinceParam'
        - $ref: '#/components/parameters/UntilParam'
      responses:
        '200':
          description: List of battles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /characters/{id}/battles:
    get:
      tags:
        - Battles
      summary: List battles for a character
      description: Returns battles where the specified character participated
      operationId: listCharacterBattles
      parameters:
        - name: id
          in: path
          required: true
          description: Character ID
          schema:
            type: string
            pattern: '^\d+$'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/SpaceTypeParam'
        - $ref: '#/components/parameters/SinceParam'
        - $ref: '#/components/parameters/UntilParam'
      responses:
        '200':
          description: List of battles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BattleListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      description: Pagination cursor from previous response
      schema:
        type: string
        format: byte
    SpaceTypeParam:
      name: spaceType
      in: query
      description: Filter by space type
      schema:
        $ref: '#/components/schemas/SpaceType'
    SystemIdParam:
      name: systemId
      in: query
      description: Filter by solar system ID
      schema:
        type: string
        pattern: '^\d+$'
    AllianceIdParam:
      name: allianceId
      in: query
      description: Filter by alliance ID
      schema:
        type: string
        pattern: '^\d+$'
    CorpIdParam:
      name: corpId
      in: query
      description: Filter by corporation ID
      schema:
        type: string
        pattern: '^\d+$'
    CharacterIdParam:
      name: characterId
      in: query
      description: Filter by character ID
      schema:
        type: string
        pattern: '^\d+$'
    SinceParam:
      name: since
      in: query
      description: Filter battles starting after this timestamp
      schema:
        type: string
        format: date-time
    UntilParam:
      name: until
      in: query
      description: Filter battles ending before this timestamp
      schema:
        type: string
        format: date-time
  schemas:
    SpaceType:
      type: string
      enum:
        - kspace
        - jspace
        - pochven
      description: |
        EVE Online space type:
        - `kspace`: Known space (empire and null-sec)
        - `jspace`: Wormhole space (J-space)
        - `pochven`: Pochven (Triglavian space)
    BattleSummary:
      type: object
      required:
        - id
        - systemId
        - systemName
        - spaceType
        - startTime
        - endTime
        - totalKills
        - totalIskDestroyed
        - zkillRelatedUrl
      properties:
        id:
          type: string
          format: uuid
          description: Unique battle identifier
        systemId:
          type: string
          description: Solar system ID (bigint as string)
        systemName:
          type: string
          nullable: true
          description: Solar system name resolved via ESI
        spaceType:
          $ref: '#/components/schemas/SpaceType'
        startTime:
          type: string
          format: date-time
          description: Timestamp of first killmail in battle
        endTime:
          type: string
          format: date-time
          description: Timestamp of last killmail in battle
        totalKills:
          type: string
          description: Total number of killmails in battle (bigint as string)
        totalIskDestroyed:
          type: string
          description: Total ISK value destroyed (bigint as string)
        zkillRelatedUrl:
          type: string
          format: uri
          description: zKillboard related kills URL
    BattleDetail:
      allOf:
        - $ref: '#/components/schemas/BattleSummary'
        - type: object
          required:
            - createdAt
            - killmails
            - participants
          properties:
            createdAt:
              type: string
              format: date-time
              description: When this battle record was created
            killmails:
              type: array
              items:
                $ref: '#/components/schemas/KillmailDetail'
            participants:
              type: array
              items:
                $ref: '#/components/schemas/BattleParticipant'
    KillmailDetail:
      type: object
      required:
        - killmailId
        - occurredAt
        - victimAllianceId
        - victimAllianceName
        - victimCorpId
        - victimCorpName
        - victimCharacterId
        - victimCharacterName
        - attackerAllianceIds
        - attackerAllianceNames
        - attackerCorpIds
        - attackerCorpNames
        - attackerCharacterIds
        - attackerCharacterNames
        - iskValue
        - zkbUrl
        - enrichment
      properties:
        killmailId:
          type: string
          description: Killmail ID (bigint as string)
        occurredAt:
          type: string
          format: date-time
        victimAllianceId:
          type: string
          nullable: true
          description: Victim's alliance ID
        victimAllianceName:
          type: string
          nullable: true
          description: Victim's alliance name
        victimCorpId:
          type: string
          nullable: true
          description: Victim's corporation ID
        victimCorpName:
          type: string
          nullable: true
          description: Victim's corporation name
        victimCharacterId:
          type: string
          nullable: true
          description: Victim's character ID
        victimCharacterName:
          type: string
          nullable: true
          description: Victim's character name
        attackerAllianceIds:
          type: array
          items:
            type: string
          description: Array of attacker alliance IDs
        attackerAllianceNames:
          type: array
          items:
            type: string
            nullable: true
          description: Array of attacker alliance names
        attackerCorpIds:
          type: array
          items:
            type: string
          description: Array of attacker corporation IDs
        attackerCorpNames:
          type: array
          items:
            type: string
            nullable: true
          description: Array of attacker corporation names
        attackerCharacterIds:
          type: array
          items:
            type: string
          description: Array of attacker character IDs
        attackerCharacterNames:
          type: array
          items:
            type: string
            nullable: true
          description: Array of attacker character names
        iskValue:
          type: string
          nullable: true
          description: ISK value of destroyed assets (bigint as string)
        zkbUrl:
          type: string
          format: uri
          description: zKillboard killmail URL
        enrichment:
          $ref: '#/components/schemas/KillmailEnrichment'
    KillmailEnrichment:
      type: object
      nullable: true
      required:
        - status
        - payload
        - error
        - fetchedAt
        - updatedAt
        - createdAt
      properties:
        status:
          type: string
          enum:
            - pending
            - processing
            - succeeded
            - failed
          description: Enrichment worker status
        payload:
          type: object
          nullable: true
          additionalProperties: true
          description: Enrichment data payload
        error:
          type: string
          nullable: true
          description: Error message if enrichment failed
        fetchedAt:
          type: string
          format: date-time
          nullable: true
          description: When enrichment data was fetched
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
    BattleParticipant:
      type: object
      required:
        - battleId
        - characterId
        - characterName
        - allianceId
        - allianceName
        - corpId
        - corpName
        - shipTypeId
        - shipTypeName
        - sideId
        - isVictim
      properties:
        battleId:
          type: string
          format: uuid
        characterId:
          type: string
          description: Character ID (bigint as string)
        characterName:
          type: string
          nullable: true
          description: Character name
        allianceId:
          type: string
          nullable: true
          description: Alliance ID (bigint as string)
        allianceName:
          type: string
          nullable: true
          description: Alliance name
        corpId:
          type: string
          nullable: true
          description: Corporation ID (bigint as string)
        corpName:
          type: string
          nullable: true
          description: Corporation name
        shipTypeId:
          type: string
          nullable: true
          description: Ship type ID (bigint as string)
        shipTypeName:
          type: string
          nullable: true
          description: Ship type name
        sideId:
          type: string
          nullable: true
          description: Battle side identifier
        isVictim:
          type: boolean
          description: Whether this participant lost their ship
    BattleListResponse:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BattleSummary'
        nextCursor:
          type: string
          nullable: true
          description: Cursor for next page of results
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'